+++ module/actors/NpcActorSheet.js
@@ -6,7 +6,7 @@ export class NpcActorSheet extends SwnActor {
   /** @inheritdoc */
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
-      classes: ["swn", "sheet", "actor"],
+      classes: ["swn", "sheet", "npc"],
       template: "systems/swn/templates/actor-sheets/npc-actor-sheet.hbs",
       width: 600,
       height: 600,
@@ -30,6 +30,27 @@ export class NpcActorSheet extends SwnActor {
     const data = super.getData()
             
     data.actorData = data.data.data
+    
+    data.actorData.equipment = []
+    data.actorData.focus = []
+    data.actorData.psionic = []
+    
+    // Sort the types of items for better readability
+    if(data.data.items){
+      for(let key in data.data.items){
+        if(data.data.items[key].type === "equipment"){
+          data.actorData.equipment.push(data.data.items[key])
+        }
+        
+        if(data.data.items[key].type === "focus"){
+          data.actorData.focus.push(data.data.items[key])
+        }
+        
+        if(data.data.items[key].type === "psionic"){
+          data.actorData.psionic.push(data.data.items[key])
+        }
+      }
+    }
     
     return data
   }